# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Corneteiro API
  version: 1.0.0
  description: |
    API para consultar dados do Cartola, analisar atletas, acompanhar pontuacoes
    e obter recomendacoes de escalação.
servers:
  - url: http://localhost:5000
    description: Ambiente local
...
paths:
...
  /recomendacoes:
    get:
      tags: [Recomendacoes]
      summary: Endpoint agregador de recomendacoes
      parameters:
        - in: query
          name: criterio
          required: true
          schema:
            type: string
            enum: [custo_beneficio, destaques_rodada, misto]
        - in: query
          name: posicao_id
          required: false
          schema:
            type: integer
        - in: query
          name: limite
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - in: query
          name: rodadas
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 38
        - in: query
          name: rodada
          required: false
          schema:
            type: integer
        - in: query
          name: ordenar_por
          required: false
          schema:
            type: string
            enum: [pontuacao_cartola, pontuacao_calculada]
            default: pontuacao_cartola
      responses:
        "200":
          description: Ranking de recomendacoes conforme o criterio informado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RecomendacaoCustoBeneficioResponse"
                  - $ref: "#/components/schemas/RecomendacaoDestaquesResponse"
                  - $ref: "#/components/schemas/RecomendacaoMistaResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /recomendacoes/custo-beneficio:
    get:
      tags: [Recomendacoes]
      summary: Endpoint legado de recomendacao por custo-beneficio
      parameters:
        - in: query
          name: posicao_id
          required: false
          schema:
            type: integer
        - in: query
          name: limite
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Ranking por custo-beneficio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecomendacaoCustoBeneficioResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /recomendacoes/destaques-rodada:
    get:
      tags: [Recomendacoes]
      summary: Endpoint legado de destaques da rodada
      parameters:
        - in: query
          name: rodada
          required: false
          schema:
            type: integer
        - in: query
          name: posicao_id
          required: false
          schema:
            type: integer
        - in: query
          name: limite
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - in: query
          name: ordenar_por
          required: false
          schema:
            type: string
            enum: [pontuacao_cartola, pontuacao_calculada]
            default: pontuacao_cartola
      responses:
        "200":
          description: Ranking de destaques da rodada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecomendacaoDestaquesResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
components:
  responses:
    ErrorResponse:
      description: Resposta de erro
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        erro:
          type: string
        detalhe:
          type: string
      required:
        - erro
...
    RecomendacaoCustoBeneficioItem:
      type: object
      properties:
        atleta_id:
          type: integer
        apelido:
          type: string
          nullable: true
        clube_id:
          type: integer
          nullable: true
        posicao_id:
          type: integer
          nullable: true
        preco_num:
          type: number
        media_num:
          type: number
        indice_custo_beneficio:
          type: number
    RecomendacaoCustoBeneficioResponse:
      type: object
      properties:
        criterio:
          type: string
          enum: [custo_beneficio]
        posicao_id:
          type: integer
          nullable: true
        limite:
          type: integer
        quantidade:
          type: integer
        recomendacoes:
          type: array
          items:
            $ref: "#/components/schemas/RecomendacaoCustoBeneficioItem"
    RecomendacaoDestaquesItem:
      type: object
      properties:
        atleta_id:
          type: integer
        apelido:
          type: string
          nullable: true
        foto:
          type: string
          nullable: true
        clube_id:
          type: integer
          nullable: true
        posicao_id:
          type: integer
          nullable: true
        entrou_em_campo:
          type: boolean
          nullable: true
        pontuacao_cartola:
          type: number
          nullable: true
        pontuacao_calculada:
          type: number
          nullable: true
        scout:
          type: object
          additionalProperties: true
    RecomendacaoDestaquesResponse:
      type: object
      properties:
        criterio:
          type: string
          enum: [destaques_rodada]
        rodada:
          oneOf:
            - type: integer
            - type: string
        posicao_id:
          type: integer
          nullable: true
        ordenar_por:
          type: string
        limite:
          type: integer
        quantidade:
          type: integer
        recomendacoes:
          type: array
          items:
            $ref: "#/components/schemas/RecomendacaoDestaquesItem"
    RecomendacaoMistaItem:
      type: object
      properties:
        atleta_id:
          type: integer
        apelido:
          type: string
          nullable: true
        clube_id:
          type: integer
          nullable: true
        posicao_id:
          type: integer
          nullable: true
        preco_num:
          type: number
        media_num:
          type: number
        partidas_validas:
          type: integer
        ultimas_pontuacoes:
          type: array
          items:
            type: number
        score_fase_recente:
          type: number
        score_consistencia:
          type: number
        score_custo_beneficio:
          type: number
        penalizacao_amostra:
          type: number
        score_recomendacao:
          type: number
    RecomendacaoMistaResponse:
      type: object
      properties:
        criterio:
          type: string
          enum: [misto]
        perfil:
          type: string
          enum: [equilibrado]
        rodadas:
          type: integer
        posicao_id:
          type: integer
          nullable: true
        limite:
          type: integer
        quantidade:
          type: integer
        recomendacoes:
          type: array
          items:
            $ref: "#/components/schemas/RecomendacaoMistaItem"